feat: Allow receiving multiple object types
For the ability to send on multiple object types to the step function,
we need to support more than the submission object. This adds support
for multiple types.

Concretely, it adds parsing for the esfTaak. The step function will
receive the submission unchanged, the taak will be passed on in an
object that has at least the pdf, attachments and reference key as well.

For determining object type, an SSM parameter is added: This contains
the supported object types, in a format like this:

`<objectType>##<objectTypeUrl>;<objectType>##<objectTypeUrl>`

This only supports for <objectType> values from
`ObjectParser.supportedObjectTypes`. All unsupported objects are
ignored.